+++ hack/tests/test-subcommand.sh
@@ -7,7 +7,7 @@ pushd test/test-framework
 # test framework with defaults
 operator-sdk test local .
 # test operator-sdk test flags
-operator-sdk test local . --global-manifest deploy/crds/cache_v1alpha1_memcached_crd.yaml --namespaced-manifest deploy/namespace-init.yaml --go-test-flags "-parallel 1" --kubeconfig $HOME/.kube/config
+operator-sdk test local . --global-manifest deploy/crds/cache_v1alpha1_memcached_crd.yaml --namespaced-manifest deploy/namespace-init.yaml --go-test-flags "-parallel 1" --kubeconfig $HOME/.kube/config --image=quay.io/coreos/operator-sdk-dev:test-framework-operator-runtime
 # test operator-sdk test local single namespace mode
 kubectl create namespace test-memcached
 # we use the test-memcached namespace for all future tests, so we only need to set this trap once